Cleaning up test integrals, making them easier to access among all tests.
