+++ torchpathdiffeq/__init__.py
@@ -2,4 +2,5 @@ from .path_integral import ode_path_integral
 from .methods import UNIFORM_METHODS, VARIABLE_METHODS
 from .base import steps, IntegralOutput
 from .serial_solver import SerialAdaptiveStepsizeSolver
-from .runge_kutta import get_parallel_RK_solver, RKParallelUniformAdaptiveStepsizeSolver, RKParallelVariableAdaptiveStepsizeSolver 
+from .runge_kutta import get_parallel_RK_solver, RKParallelUniformAdaptiveStepsizeSolver, RKParallelVariableAdaptiveStepsizeSolver
+from .examples import ODE_dict
